feat: add entities infrastrucuture configurations

+++ src/Domain/Feature/FeatureEntity.cs
@@ -5,9 +5,9 @@ using Domain.Power;
 
 namespace Domain.Feature
 {
-    public class FeatureEntity : BaseAuditableEntity<uint>
+    public class FeatureEntity : BaseAuditableEntity<ushort>
     {
-        public uint GameId { get; set; } = default;
+        public ushort GameId { get; set; } = default;
         public string Name { get; set; } = string.Empty;
         public GameEntity Game { get; set; } = null!;
         public List<PowerEntity> Powers { get; set; } = [];
