+++ src/Domain/Power/PowerEntity.cs
@@ -4,9 +4,9 @@ using Domain.Feature;
 
 namespace Domain.Power
 {
-    public class PowerEntity : BaseAuditableEntity<uint>
+    public class PowerEntity : BaseAuditableEntity<ushort>
     {
-        public uint CardId { get; set; } = default;
+        public ushort CardId { get; set; } = default;
         public uint FeatureId { get; set; } = default;
         public uint Value { get; set; } = default;
         public CardEntity Card { get; set; } = null!;
